support for using the mocks from other modules
